feedback + fixing unit tests
