+++ integration-tests/basic-test-app/app/(tabs)/_layout.tsx
@@ -46,13 +46,6 @@ export default function TabLayout() {
           tabBarAccessibilityLabel: "MISC TESTING",
         }}
       />
-      <Tabs.Screen
-        name="render"
-        options={{
-          title: "render",
-          tabBarAccessibilityLabel: "RENDER TESTING",
-        }}
-      />
     </Tabs>
   );
 }
