--- integration-tests/basic-test-app/app/(tabs)/render.ts
@@ -1,2 +1,0 @@
-import {RenderTestingScreen} from "@embrace-io/react-native-test-harness";
-export default RenderTestingScreen;
