+++ integration-tests/basic-test-app/metro.config.js
@@ -4,13 +4,4 @@ const {getDefaultConfig} = require("expo/metro-config");
 /** @type {import('expo/metro-config').MetroConfig} */
 const config = getDefaultConfig(__dirname);
 
-config.transformer.minifierConfig = {
-  keep_classnames: true,
-  keep_fnames: true,
-  mangle: {
-    keep_classnames: true,
-    keep_fnames: true,
-  },
-};
-
 module.exports = config;
