+++ integration-tests/test-harness/src/index.ts
@@ -5,4 +5,3 @@ export * from "./screens/OTLPTestingScreen";
 export * from "./screens/PropertyTestingScreen";
 export * from "./screens/SpanTestingScreen";
 export * from "./screens/TracerProviderTestingScreen";
-export * from "./screens/RenderTestingScreen";
